requirements.txt
flask>=3.0.0
proxmoxer>=2.0.1
requests>=2.28.0
urllib3>=2.0.0
pystray>=0.19.0
Pillow>=9.0.0
